scripts/requirements.txt
requests>=2.25.0
pyjwt[crypto]>=2.4.0